# Increase the aggregation buffer sizes for non-ugni configurations