@mhmerrill 2021-04-05
@ronawho so at 128cores/node and 16 nodes this would be 256MB in buffers (512MB at 32 nodes)
cores*8192*(8addr+8data)*nodes = 128*8192*16*16 = 2**(7+13+4+4) =  2**28 bytes = 256MB per aggregator