@mhmerrill 2021-04-05
We should set an upper limit and divide it by the number of cores time the number of locales or something like that so we don't blow the bank at higher node counts...