@ronawho 2021-04-06
> `256MB` per aggregator

I think that's the per node memory usage, not per aggregator. I agree the memory usage can be a problem, but that's been a problem since the introduction of aggregators and this just doubles the value for some platforms. My general assumption is that as numCores increases so does memory per node, so I don't think 128 core systems changes what percentage of memory we're using too much.

Longer term I expect to use multi-hop aggregators to reduce memory overheads.

> We should set an upper limit and divide it by the number of cores time the number of locales or something like that so we don't blow the bank at higher node counts

Note that even at ~256 nodes I was seeing perf improvements from increasing the buffer size so doing that would limit performance at scale. 256 nodes with 128 cores each would be 32 MiB per aggregator, 4GiB per node which isn't great but I don't this is bad.